Moved `checks` out of `planning.rs`
